refactor: DI 패키지 내부 객체명 변경
