+++ app/src/main/java/woowacourse/shopping/di/DatabaseInjection.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.di
 import android.content.Context
 import woowacourse.shopping.data.database.ShoppingDatabase
 
-object DatabaseModule {
+object DatabaseInjection {
     private var _database: ShoppingDatabase? = null
     val database: ShoppingDatabase get() = _database ?: throw IllegalStateException()
