+++ app/src/main/java/woowacourse/shopping/di/PreferenceInjection.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import woowacourse.shopping.data.preference.AuthSharedPreference
 
 @SuppressLint("StaticFieldLeak")
-object PreferenceModule {
+object PreferenceInjection {
     private var _authSharedPreference: AuthSharedPreference? = null
     val authSharedPreference: AuthSharedPreference get() = _authSharedPreference ?: throw IllegalStateException()
