+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -8,12 +8,12 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.CreationExtras
 import kotlinx.coroutines.launch
-import woowacourse.shopping.di.UseCaseModule.decreaseCartProductQuantityUseCase
-import woowacourse.shopping.di.UseCaseModule.getCartProductsUseCase
-import woowacourse.shopping.di.UseCaseModule.getCartRecommendProductsUseCase
-import woowacourse.shopping.di.UseCaseModule.getCatalogProductUseCase
-import woowacourse.shopping.di.UseCaseModule.increaseCartProductQuantityUseCase
-import woowacourse.shopping.di.UseCaseModule.removeCartProductUseCase
+import woowacourse.shopping.di.UseCaseInjection.decreaseCartProductQuantityUseCase
+import woowacourse.shopping.di.UseCaseInjection.getCartProductsUseCase
+import woowacourse.shopping.di.UseCaseInjection.getCartRecommendProductsUseCase
+import woowacourse.shopping.di.UseCaseInjection.getCatalogProductUseCase
+import woowacourse.shopping.di.UseCaseInjection.increaseCartProductQuantityUseCase
+import woowacourse.shopping.di.UseCaseInjection.removeCartProductUseCase
 import woowacourse.shopping.domain.model.Page
 import woowacourse.shopping.domain.model.Page.Companion.EMPTY_PAGE
 import woowacourse.shopping.domain.model.Product.Companion.MINIMUM_QUANTITY
