+++ app/src/main/java/woowacourse/shopping/ui/catalog/CatalogViewModel.kt
@@ -8,13 +8,13 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.CreationExtras
 import kotlinx.coroutines.launch
-import woowacourse.shopping.di.UseCaseModule.decreaseCartProductQuantityUseCase
-import woowacourse.shopping.di.UseCaseModule.getCartProductsQuantityUseCase
-import woowacourse.shopping.di.UseCaseModule.getCatalogProductUseCase
-import woowacourse.shopping.di.UseCaseModule.getCatalogProductsByProductIdsUseCase
-import woowacourse.shopping.di.UseCaseModule.getCatalogProductsUseCase
-import woowacourse.shopping.di.UseCaseModule.getSearchHistoryUseCase
-import woowacourse.shopping.di.UseCaseModule.increaseCartProductQuantityUseCase
+import woowacourse.shopping.di.UseCaseInjection.decreaseCartProductQuantityUseCase
+import woowacourse.shopping.di.UseCaseInjection.getCartProductsQuantityUseCase
+import woowacourse.shopping.di.UseCaseInjection.getCatalogProductUseCase
+import woowacourse.shopping.di.UseCaseInjection.getCatalogProductsByProductIdsUseCase
+import woowacourse.shopping.di.UseCaseInjection.getCatalogProductsUseCase
+import woowacourse.shopping.di.UseCaseInjection.getSearchHistoryUseCase
+import woowacourse.shopping.di.UseCaseInjection.increaseCartProductQuantityUseCase
 import woowacourse.shopping.domain.model.Page.Companion.UNINITIALIZED_PAGE
 import woowacourse.shopping.domain.usecase.DecreaseCartProductQuantityUseCase
 import woowacourse.shopping.domain.usecase.GetCartProductsQuantityUseCase
